requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1

# AI/ML
langchain==0.0.330
openai==1.3.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
python-dateutil==2.8.2
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1